+++ src/scripts/public.js
@@ -1,4 +1,5 @@
 import '../styles/public.scss';
+import 'material-symbols/outlined.css';
 
 /**
  * All the scripting for the frontend goes here.